Step size can now be double to match with the network latency
Changed step size to double and ability to match step size to latency value

+++ SimulatorBridger-traffic_information_collector/src/main/java/uk/ncl/giacomobergami/traffic_converter/abstracted/TrafficConverter.java
@@ -151,6 +151,6 @@ public abstract class TrafficConverter {
         return true;
     }
 
-    public abstract boolean runSimulator(long begin, long end, long step);
+    public abstract boolean runSimulator(long begin, long end, double step);
 
 }
